+++ M-M_Sortor/M-M_Sortor/Stepper.h
@@ -9,6 +9,6 @@
 #ifndef STEPPER_H_
 #define STEPPER_H_
 void initStepper();
-void step(unsigned int steps, unsigned int direction);
+void step(float steps, unsigned int direction);
 
 #endif /* STEPPER_H_ */
